Sprint 11.1: Session state model + ManifestDraft + specta bindings
- crates/ir/src/session.rs (NEW): Full session state machine
  * 10 types: WorkflowType, FileEntry, ManifestRef, ProgressState,
    ReviewItemStatus, ReviewItem, WorkflowResult, ExportSummary,
    DetectedSection, SessionState (7 variants)
  * 12 pure transition functions covering all 13 state transitions
  * detect_workflow_hint() for advisory SPEC_/DWG_/SUB_ prefix detection
  * 17 unit tests passing — ManifestDraft path fully covered
  * cfg_attr(feature='specta') derives on all IPC-boundary types
- crates/ir/Cargo.toml: optional specta feature flag (=2.0.0-rc.22)
- crates/ir/src/types.rs: BBox gains specta::Type derive (feature-gated)
- crates/ir/src/lib.rs: pub mod session; + full re-export of all session types

- apps/desktop-gui/src-tauri/Cargo.toml: added conset-pdf-ir (specta feature),
  specta =rc.22, tauri-specta =rc.21, specta-typescript =0.0.9
- apps/desktop-gui/src-tauri/src/lib.rs: specta_builder() fn; auto-export
  bindings on #[cfg(debug_assertions)]; single generate_handler! for all cmds
- apps/desktop-gui/src-tauri/src/commands.rs: #[specta::specta] on the 3
  typed-return commands; specta::Type derives on FileFilter + ManifestValidationResult;
  sections_targeted changed usize -> u32 (BigInt-safe)
- apps/desktop-gui/src-tauri/src/bin/gen_bindings.rs (NEW): standalone
  gen_bindings binary for CI / pre-commit hooks

DoD: 2 (17 session transition tests), 2b (bindings.ts generated) -> both ✅
cargo check --workspace: EXIT 0
cargo test -p conset-pdf-ir -p conset-pdf-desktop-gui-tauri: 72/72 pass

+++ apps/desktop-gui/src-tauri/src/commands.rs
@@ -14,17 +14,17 @@ use crate::backend_process;
 // Supplementary types for commands not covered by WorkflowRequest/Response
 // ---------------------------------------------------------------------------
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
 pub struct FileFilter {
     pub name: String,
     pub extensions: Vec<String>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, specta::Type)]
 pub struct ManifestValidationResult {
     pub valid: bool,
     pub errors: Vec<String>,
-    pub sections_targeted: usize,
+    pub sections_targeted: u32,
 }
 
 // ---------------------------------------------------------------------------
@@ -191,6 +191,7 @@ pub fn cmd_visualize(
 
 /// Stub — open-file-dialog; real implementation wires tauri-plugin-dialog on the TS side.
 /// This command exists so the TypeScript invoke surface is complete and typed.
+#[specta::specta]
 #[tauri::command]
 pub fn cmd_open_file_dialog(
     _title: String,
@@ -202,6 +203,7 @@ pub fn cmd_open_file_dialog(
 }
 
 /// Stub — save-file-dialog; real implementation wires tauri-plugin-dialog on the TS side.
+#[specta::specta]
 #[tauri::command]
 pub fn cmd_save_file_dialog(
     _title: String,
@@ -219,6 +221,7 @@ pub fn cmd_save_file_dialog(
 /// Checks that all `section_id` values exist in the detected segment index and
 /// that operation types / required fields are present and well-formed.
 /// Lives here (not in engine) because it only parses JSON — no PDFium needed.
+#[specta::specta]
 #[tauri::command]
 pub fn cmd_validate_manifest(
     manifest_json: String,
@@ -238,7 +241,7 @@ fn validate_manifest_inner(
     let index: Value = serde_json::from_str(segment_index_json)?;
 
     let mut errors: Vec<String> = Vec::new();
-    let mut sections_targeted: usize = 0;
+    let mut sections_targeted: u32 = 0;
 
     // Collect known section IDs from the segment index.
     let known_ids: std::collections::HashSet<String> = index
